Refactor responsibilities based on review feedback.
